scripts/validate_data_integrity.py
#!/usr/bin/env python3
"""
Multi-Dataset Integrity Validation for Medical Imaging Pipeline

Comprehensive validation of data quality across all datasets:
- Image file integrity and format consistency
- Class distribution analysis and imbalance assessment
- Cross-dataset duplicate detection  
- Clinical annotation validation
- Processing pipeline verification

Essential for medical AI projects to ensure data quality and regulatory compliance.
"""

import logging
from pathlib import Path
from typing import Dict, List, Tuple, Set, Optional
import hashlib
import json
from collections import defaultdict, Counter

import pandas as pd
import numpy as np
from PIL import Image
from omegaconf import DictConfig, OmegaConf
import hydra

logger = logging.getLogger(__name__)

@hydra.main(version_base=None, config_path="../config", config_name="dataset_config")
def validate_data_integrity(cfg: DictConfig) -> None:
    """
    Comprehensive data integrity validation across all datasets.
    
    Args:
        cfg: Dataset configuration from YAML
        
    Validation Steps:
    1. File system integrity (paths, formats, accessibility)
    2. Image quality validation (corruption, resolution, format)
    3. Class distribution analysis
    4. Cross-dataset duplicate detection
    5. Clinical annotation consistency check
    6. Processing pipeline validation
    """
    
    logger.info("Starting comprehensive data integrity validation")
    
    # Validation results storage
    validation_results = {
        'file_system': {},
        'image_quality': {},
        'class_distribution': {},
        'duplicates': {},
        'clinical_annotations': {},
        'processing_pipeline': {}
    }
    
    # Run validation modules
    validation_results['file_system'] = _validate_file_system(cfg)
    validation_results['image_quality'] = _validate_image_quality(cfg)  
    validation_results['class_distribution'] = _analyze_class_distributions(cfg)
    validation_results['duplicates'] = _detect_cross_dataset_duplicates(cfg)
    validation_results['clinical_annotations'] = _validate_clinical_annotations(cfg)
    validation_results['processing_pipeline'] = _validate_processing_pipeline(cfg)
    
    # Generate comprehensive report
    _generate_validation_report(validation_results)
    
    logger.info("Data integrity validation completed")


def _validate_file_system(cfg: DictConfig) -> Dict:
    """
    Validate file system integrity and accessibility.
    
    Checks:
    - All dataset paths exist and are accessible
    - Expected directory structure is present
    - File counts match expectations
    - Permissions are correct for processing
    
    Returns:
        Dict with file system validation results
    """
    logger.info("Validating file system integrity...")
    
    results = {
        'paths_exist': {},
        'directory_structure': {},
        'file_counts': {},
        'permissions': {},
        'errors': []
    }
    
    for dataset_name, dataset_cfg in cfg.datasets.items():
        try:
            dataset_path = Path(dataset_cfg.path)
            results['paths_exist'][dataset_name] = dataset_path.exists()
            
            if dataset_path.exists():
                # TODO: Implement detailed file system checks
                # Check directory structure
                # Count files by type
                # Verify read/write permissions
                # Validate expected class directories exist
                
                results['file_counts'][dataset_name] = _count_files_by_extension(dataset_path)
                results['directory_structure'][dataset_name] = _validate_directory_structure(dataset_path, dataset_cfg)
                
            else:
                results['errors'].append(f"Dataset path not found: {dataset_path}")
                
        except Exception as e:
            results['errors'].append(f"Error validating {dataset_name}: {str(e)}")
    
    return results


def _validate_image_quality(cfg: DictConfig) -> Dict:
    """
    Validate image file integrity and quality metrics.
    
    Quality checks:
    - File corruption detection
    - Resolution consistency
    - Format validation  
    - Color space verification
    - Metadata extraction
    
    Returns:
        Dict with image quality validation results
    """
    logger.info("Validating image quality and integrity...")
    
    results = {
        'corrupted_images': [],
        'resolution_stats': {},
        'format_distribution': {},
        'quality_metrics': {},
        'errors': []
    }
    
    for dataset_name, dataset_cfg in cfg.datasets.items():
        if dataset_name == "sumotosima_github":
            continue  # Skip CSV dataset
            
        try:
            dataset_path = Path(dataset_cfg.path)
            if not dataset_path.exists():
                continue
                
            # TODO: Implement comprehensive image quality validation
            # Scan all image files in dataset
            # Check for corruption using PIL
            # Extract resolution and format information
            # Calculate quality metrics (sharpness, contrast, etc.)
            # Detect outliers in image properties
            
            results['format_distribution'][dataset_name] = _analyze_image_formats(dataset_path)
            results['resolution_stats'][dataset_name] = _analyze_image_resolutions(dataset_path)
            
        except Exception as e:
            results['errors'].append(f"Error validating images in {dataset_name}: {str(e)}")
    
    return results


def _analyze_class_distributions(cfg: DictConfig) -> Dict:
    """
    Analyze class distributions across all datasets.
    
    Analysis:
    - Class counts per dataset
    - Imbalance severity assessment
    - Cross-dataset class overlap
    - Augmentation requirements calculation
    
    Returns:
        Dict with class distribution analysis
    """
    logger.info("Analyzing class distributions...")
    
    results = {
        'per_dataset_distributions': {},
        'combined_distribution': {},
        'imbalance_metrics': {},
        'augmentation_requirements': {},
        'class_mapping_issues': []
    }
    
    for dataset_name, dataset_cfg in cfg.datasets.items():
        if dataset_name == "sumotosima_github":
            continue
            
        # TODO: Implement class distribution analysis
        # Count images per class in each dataset
        # Calculate imbalance ratios
        # Identify severely underrepresented classes
        # Recommend augmentation factors
        # Check for class mapping consistency
        
        results['per_dataset_distributions'][dataset_name] = _count_classes_in_dataset(
            Path(dataset_cfg.path)
        )
    
    return results


def _detect_cross_dataset_duplicates(cfg: DictConfig) -> Dict:
    """
    Detect duplicate images across datasets using perceptual hashing.
    
    Important for medical imaging to avoid data leakage between train/test sets.
    
    Detection methods:
    - File hash comparison (exact duplicates)
    - Perceptual hashing (near duplicates)
    - Metadata comparison
    - Visual similarity detection
    
    Returns:
        Dict with duplicate detection results
    """
    logger.info("Detecting cross-dataset duplicates...")
    
    results = {
        'exact_duplicates': [],
        'near_duplicates': [],
        'duplicate_pairs': [],
        'statistics': {}
    }
    
    # TODO: Implement duplicate detection
    # Calculate file hashes for exact duplicate detection
    # Use perceptual hashing (pHash) for near-duplicate detection
    # Compare images across different datasets
    # Generate duplicate pair reports
    # Calculate duplicate statistics
    
    return results


def _validate_clinical_annotations(cfg: DictConfig) -> Dict:
    """
    Validate clinical annotations in Sumotosima dataset.
    
    Validation:
    - CSV structure integrity
    - Clinical description consistency
    - Class label validation
    - Text quality assessment
    
    Returns:
        Dict with clinical annotation validation results
    """
    logger.info("Validating clinical annotations...")
    
    results = {
        'csv_integrity': {},
        'description_quality': {},
        'class_consistency': {},
        'text_statistics': {}
    }
    
    # TODO: Implement clinical annotation validation
    # Load and validate CSV structure
    # Check for missing or malformed entries
    # Analyze clinical description quality
    # Validate consistency with image classifications
    # Generate text statistics and quality metrics
    
    return results


def _validate_processing_pipeline(cfg: DictConfig) -> Dict:
    """
    Validate the image processing pipeline functionality.
    
    Tests:
    - CLAHE enhancement quality
    - Resize operation accuracy
    - Format conversion integrity
    - Processing speed benchmarks
    
    Returns:
        Dict with processing pipeline validation results
    """
    logger.info("Validating processing pipeline...")
    
    results = {
        'preprocessing_quality': {},
        'performance_benchmarks': {},
        'output_consistency': {},
        'errors': []
    }
    
    # TODO: Implement processing pipeline validation
    # Test preprocessing functions on sample images
    # Validate CLAHE enhancement quality
    # Check resize operation accuracy
    # Benchmark processing speed
    # Verify output format consistency
    
    return results


# Helper functions (stubs for actual implementation)

def _count_files_by_extension(path: Path) -> Dict[str, int]:
    """Count files by extension in directory."""
    # TODO: Implement file counting
    return {}

def _validate_directory_structure(path: Path, cfg: DictConfig) -> Dict:
    """Validate expected directory structure."""
    # TODO: Implement structure validation
    return {}

def _analyze_image_formats(path: Path) -> Dict:
    """Analyze distribution of image formats."""
    # TODO: Implement format analysis
    return {}

def _analyze_image_resolutions(path: Path) -> Dict:
    """Analyze image resolution statistics."""
    # TODO: Implement resolution analysis
    return {}

def _count_classes_in_dataset(path: Path) -> Dict:
    """Count images per class in dataset."""
    # TODO: Implement class counting
    return {}

def _generate_validation_report(validation_results: Dict) -> None:
    """Generate comprehensive validation report."""
    logger.info("Generating validation report...")
    
    # TODO: Implement report generation
    # Create detailed HTML/PDF report
    # Include visualizations and statistics
    # Highlight critical issues and recommendations
    # Save report to outputs directory
    
    logger.info("✓ Validation report generated")


if __name__ == "__main__":
    validate_data_integrity()


# Usage:
# python scripts/validate_data_integrity.py
# python scripts/validate_data_integrity.py --config-dir=../config --config-name=dataset_config